Edited pull request #287 after 2nd review
