+++ app/models/course.rb
@@ -5,8 +5,9 @@ class Course < ActiveRecord::Base
   default_scope { order(number: :asc) }
   after_save :send_notification
 
+
   def send_notification
-    EventResponders::CourseResponder.new.call(self)
+    ::CoursesResponder.new.call(self)
   end
 
   def self.get code, number
