+++ app/models/section.rb
@@ -7,8 +7,9 @@ class Section < ActiveRecord::Base
   after_save :update_conflicts!, if: :periods_changed?
   after_save :send_notification
 
+
   def send_notification
-     EventResponders::SectionResponder.new.call(self)
+     ::SectionsResponder.new.call(self)
   end
