--- app/responders/core_responder.rb
@@ -1,18 +1,0 @@
-module EventResponders
-
-    class CourseResponder < ApplicationResponder
-       topic :course_changee, required: true
-
-        def respond(event)
-            respond_to :course_change, event.to_json(:only => [:uuid,:department_id, :name, :number, :min_credits, :max_credits, :created_at , :updated_at, :description, :tags  ])
-        end
-    end
-
-    class SectionResponder < ApplicationResponder
-       topic :section_change, required: true
-
-       def respond(event)
-           respond_to :section_change , event.to_json(:only => [:uuid,:course_id, :name, :crn, :seats, :seats, :seats_taken, :created_at, :updated_at, :num_periods, :instructors, :conflicts ])
-       end
-    end
-end
